Add comb method
