+++ fastcrypto/src/groups/multiplier/fixed_window.rs
@@ -1,0 +1,132 @@
+use crate::groups::multiplier::{integer_utils, ScalarMultiplier};
+use crate::groups::{Doubling, GroupElement};
+use crate::serde_helpers::ToFromByteArray;
+use std::mem::size_of;
+
+/// Performs scalar multiplication using a fixed window method. If the addition and doubling operations for
+/// `G` are constant time, the `mul`  is also constant time. The `CACHE_SIZE` should ideally be a power of
+/// two. The `SCALAR_SIZE` is the number of bytes in the byte representation of the scalar type `S`, and we
+/// assume that the `S::from_bytes` method returns the scalar in little-endian format.
+pub struct FixedWindowMultiplier<
+    G: GroupElement<ScalarType = S>,
+    S: GroupElement + ToFromByteArray<SCALAR_SIZE>,
+    const CACHE_SIZE: usize,
+    const SCALAR_SIZE: usize,
+> {
+    /// Precomputed multiples of the base element, B, up to (2^WINDOW_WIDTH - 1) * B.
+    cache: [G; CACHE_SIZE],
+}
+
+impl<
+        G: GroupElement<ScalarType = S>,
+        S: GroupElement + ToFromByteArray<SCALAR_SIZE>,
+        const CACHE_SIZE: usize,
+        const SCALAR_SIZE: usize,
+    > FixedWindowMultiplier<G, S, CACHE_SIZE, SCALAR_SIZE>
+{
+    /// The number of bits in the window. This is equal to the floor of the log2 of the cache size.
+    const WINDOW_WIDTH: usize = 8 * size_of::<usize>() - CACHE_SIZE.leading_zeros() as usize - 1;
+
+    /// Get the multiple `s * base_element` for a scalar `s` with `0 <= s < 2^Self::WINDOW_WIDTH`.
+    /// If `s` is not in this range the method will panic.
+    fn get_precomputed_multiple(&self, s: usize) -> G {
+        self.cache[s]
+    }
+}
+
+impl<
+        G: GroupElement<ScalarType = S> + Doubling,
+        S: GroupElement + ToFromByteArray<SCALAR_SIZE>,
+        const CACHE_SIZE: usize,
+        const SCALAR_SIZE: usize,
+    > ScalarMultiplier<G> for FixedWindowMultiplier<G, S, CACHE_SIZE, SCALAR_SIZE>
+{
+    fn new(base_element: G) -> Self {
+        // Compute cache [0, base_element, 2 * base_element, ..., (2^WINDOW_WIDTH - 1) * base_element]
+        let mut cache = [G::zero(); CACHE_SIZE];
+        for i in 1..CACHE_SIZE {
+            cache[i] = cache[i - 1] + base_element;
+        }
+        Self { cache }
+    }
+
+    fn mul(&self, scalar: &S) -> G {
+        let scalar_bytes = scalar.to_byte_array();
+
+        let limbs = integer_utils::compute_base_2w_expansion::<SCALAR_SIZE>(
+            &scalar_bytes,
+            Self::WINDOW_WIDTH,
+        );
+
+        // Number of digits in the base 2^window_size representation of the scalar
+        let mut r: G = self.get_precomputed_multiple(limbs[limbs.len() - 1]);
+
+        for i in (0..=(limbs.len() - 2)).rev() {
+            for _ in 1..=Self::WINDOW_WIDTH {
+                r = r.double();
+            }
+            r += self.get_precomputed_multiple(limbs[i]);
+        }
+        r
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::groups::ristretto255::{RistrettoPoint, RistrettoScalar};
+    use crate::groups::secp256r1::{ProjectivePoint, Scalar};
+
+    #[test]
+    fn test_scalar_multiplication_ristretto() {
+        let multiplier = FixedWindowMultiplier::<RistrettoPoint, RistrettoScalar, 16, 32>::new(
+            RistrettoPoint::generator(),
+        );
+        let scalar = RistrettoScalar::from(123456789);
+        let expected = RistrettoPoint::generator() * scalar;
+        let actual = multiplier.mul(&scalar);
+        assert_eq!(expected, actual);
+    }
+
+    #[test]
+    fn test_scalar_multiplication_secp256r1() {
+        let scalar = Scalar::from(123456789);
+        let expected = ProjectivePoint::generator() * scalar;
+
+        let multiplier = FixedWindowMultiplier::<ProjectivePoint, Scalar, 15, 32>::new(
+            ProjectivePoint::generator(),
+        );
+        let actual = multiplier.mul(&scalar);
+        assert_eq!(expected, actual);
+
+        let multiplier = FixedWindowMultiplier::<ProjectivePoint, Scalar, 16, 32>::new(
+            ProjectivePoint::generator(),
+        );
+        let actual = multiplier.mul(&scalar);
+        assert_eq!(expected, actual);
+
+        let multiplier = FixedWindowMultiplier::<ProjectivePoint, Scalar, 17, 32>::new(
+            ProjectivePoint::generator(),
+        );
+        let actual = multiplier.mul(&scalar);
+        assert_eq!(expected, actual);
+
+        let multiplier = FixedWindowMultiplier::<ProjectivePoint, Scalar, 32, 32>::new(
+            ProjectivePoint::generator(),
+        );
+        let actual = multiplier.mul(&scalar);
+        assert_eq!(expected, actual);
+
+        let multiplier = FixedWindowMultiplier::<ProjectivePoint, Scalar, 64, 32>::new(
+            ProjectivePoint::generator(),
+        );
+        let actual = multiplier.mul(&scalar);
+        assert_eq!(expected, actual);
+
+        let multiplier = FixedWindowMultiplier::<ProjectivePoint, Scalar, 512, 32>::new(
+            ProjectivePoint::generator(),
+        );
+        let actual = multiplier.mul(&scalar);
+        assert_eq!(expected, actual);
+    }
+}
